pyproject.toml
[tool.poetry]
name = "hackathon-semcomp-2024"
version = "0.1.0"
description = ""
authors = ["Antônio Marcos <antoniomarcos.souza2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
pandas = "^2.2.3"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
nicegui = "^2.7.0"
langchain-core = "^0.3.21"
langchain-community = "^0.3.8"
langchain-groq = "^0.2.1"
sqlfluff = "^3.2.5"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = []
